+++ packages/proto-gen-dts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.12](https://github.com/planjs/stan/compare/proto-gen-dts@0.7.11...proto-gen-dts@0.7.12) (2021-06-17)
+
+**Note:** Version bump only for package proto-gen-dts
+
+
+
+
+
 ## [0.7.11](https://github.com/planjs/stan/compare/proto-gen-dts@0.7.10...proto-gen-dts@0.7.11) (2021-05-25)
 
 **Note:** Version bump only for package proto-gen-dts